Stubby implementation of sceSysmoduleGetModuleInfoForUnwind
